chore[PNT] separate directories for controllers/services
