+++ src/models/tournament/controllers/tournaments.controller.ts
@@ -1,8 +1,8 @@
 import { Body, Controller, Get, Logger, Param, ParseIntPipe, Post, Put, Query, ValidationPipe } from '@nestjs/common';
 import { ApiOperation, ApiParam, ApiResponse, ApiTags } from '@nestjs/swagger';
-import { DefaultExceptionDto } from './dto/error.dto';
-import { CreateTournamentDto, TournamentDto, TournamentsList, TournamentsQuery } from './dto/tournaments.dto';
-import { TournamentsService } from './tournaments.service';
+import { DefaultExceptionDto } from '../dto/error.dto';
+import { CreateTournamentDto, TournamentDto, TournamentsList, TournamentsQuery } from '../dto/tournaments.dto';
+import { TournamentsService } from '../services/tournaments.service';
 
 @ApiTags('Tournaments')
 @Controller({
