+++ src/models/tournament/tournaments.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
-import { TournamentsController } from './tournaments.controller';
+import { TournamentsController } from './controllers/tournaments.controller';
 import { ThrottlerModule } from '@nestjs/throttler';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { TournamentsService } from './tournaments.service';
+import { TournamentsService } from './services/tournaments.service';
 import { Tournament } from './entities/tournament.entity';
 import { Game } from './entities/game.entity';
 import { GameRules } from './entities/gameRules.entity';
@@ -10,21 +10,34 @@ import { Matchup } from './entities/matchup.entity';
 import { Group } from './entities/group.entity';
 import { Stage } from './entities/stage.entity';
 import { GameService } from './services/game.service';
-import { GameController } from './game.controller';
+import { GameController } from './controllers/game.controller';
 import { SeeksService } from './services/seeks.service';
-import { SeeksController } from './seeks.controller';
+import { SeeksController } from './controllers/seeks.controller';
 import { HttpModule } from '@nestjs/axios';
-import { MatchupsController } from './matchups.controller';
+import { MatchupsController } from './controllers/matchups.controller';
 import { MatchupsService } from './services/matchups.service';
 import { GameRulesService } from './services/gameRules.service';
-import { GameRulesController } from './gameRules.controller';
+import { GameRulesController } from './controllers/gameRules.controller';
 import { ConfigService } from '@nestjs/config';
+import { StagesController } from './controllers/stages.controller';
+import { GroupsController } from './controllers/groups.controller';
+import { StagesService } from './services/stages.service';
+import { GroupsService } from './services/groups.service';
 
 @Module({
-	controllers: [TournamentsController, GameController, SeeksController, MatchupsController, GameRulesController],
+	controllers: [
+		SeeksController,
+		GameRulesController,
+		TournamentsController,
+		StagesController,
+		GroupsController,
+		MatchupsController,
+		GameController,
+	],
 	imports: [
 		HttpModule,
-		TypeOrmModule.forFeature([Tournament, Stage, Group, Matchup, Game, GameRules]),
+		TypeOrmModule.forFeature([
+			Tournament, Stage, Group, Matchup, Game, GameRules]),
 		ThrottlerModule.forRootAsync({
 			useFactory: () => ({
 				ttl: 60,
@@ -32,6 +45,15 @@ import { ConfigService } from '@nestjs/config';
 			}),
 		}),
 	],
-	providers: [TournamentsService, GameService, SeeksService, MatchupsService, GameRulesService, ConfigService],
+	providers: [
+		SeeksService,
+		GameRulesService,
+		TournamentsService,
+		StagesService,
+		GroupsService,
+		MatchupsService,
+		GameService,
+		ConfigService,
+	],
 })
-export class TournamentsModule {}
+export class TournamentsModule { }
